Merge Phase1

+++ src/models/Plant.java
@@ -1,5 +1,8 @@
 package models;
 
+/**
+ * The Plant class is responsible for holding the plant's data. All of its attributes are final.
+ */
 public class Plant {
     private final String name;
     private final String type;
@@ -16,7 +19,7 @@ public class Plant {
 
     public Plant(String name, String type, int harvestTime,
             int waterMin, int waterMax, int fertMin, int fertMax, int minProduce,
-            int maxProduce, int storePrice, int retail, int expYield)
+            int maxProduce, int storePrice, int retail, float expYield)
     {
         this.name = name;
         this.type = type;
@@ -32,51 +35,102 @@ public class Plant {
         this.expYield = expYield;
     }
 
+    /**
+     * Returns the name of the plant.
+     * @return the name of the plant.
+     */
     public String getName()
     {
         return this.name;
     }
 
+
+    /**
+     * Returns the type of the plant.
+     * @return the type of the plant.
+     */
     public String getType()
     {
         return this.type;
     }
 
+
+    /**
+     * Returns the harvest time of the plant.
+     * @return the harvest time of the plant.
+     */
     public int getHarvestTime() {
         return this.harvestTime;
     }
 
+    /**
+     * Returns the minimum water requirement of the plant.
+     * @return the minimum water requirement of the plant.
+     */
     public int getWaterMin() {
         return this.waterMin;
     }
 
+    /**
+     * Returns the maximum water requirement of the plant.
+     * @return the maximum water requirement of the plant.
+     */
     public int getWaterMax() {
         return this.waterMax;
     }
 
+    /**
+     * Returns the minimum fertilizer requirement of the plant.
+     * @return the minimum fertilizer requirement of the plant.
+     */
     public int getFertMin() {
         return this.fertMin;
     }
 
+    /**
+     * Returns the maximum fertilizer requirement of the plant.
+     * @return the maximum fertilizer requirement of the plant.
+     */
     public int getFertMax() {
         return this.fertMax;
     }
 
+    /**
+     * Returns the minimum produce of the plant.
+     * @return the minimum produce of the plant.
+     */
     public int getMinProduce(){
         return this.minProduce;
     }
+
+    /**
+     * Returns the maximum produce of the plant.
+     * @return the maximum produce of the plant.
+     */
     public int getMaxProduce() {
         return this.maxProduce;
     }
 
+    /**
+     * Returns the store price of the seed.
+     * @return the store price of the seed.
+     */
     public int getStorePrice() {
         return this.storePrice;
     }
 
+    /**
+     * Returns the retail price of the produce.
+     * @return the retail price of the produce.
+     */
     public int getRetail() {
         return this.retail;
     }
 
+    /**
+     * Returns the experience yield of the plant.
+     * @return the experience yield of the plant.
+     */
     public float getExpYield() {
         return this.expYield;
     }
